DAO e Controller

+++ ProjetoIntegradorFarmacia/src/main/java/com/mycompany/projetointegradorfarmacia/model/Produto.java
@@ -60,7 +60,7 @@ public class Produto {
         this.pVenda = pVenda;
     }
 
-    public double getQuantProd() {
+    public int getQuantProd() {
         return quantProd;
     }
